Clumsy typescript / deno rewrite
